Update listing status and break out adapter files
