+++ src/adapters/tests/leasing-adapter/mocks.ts
